YSK/requirements.txt
yandex-speechkit
